requirements.txt
requests
lxml
mysql-connector-repackaged
schedule
